refactor(config): split Configuration into 3 focused collaborators
Phase 3 of #234. The 700-line Configuration god-class is now a thin
IConfiguration facade (~105 lines) composing:
  - SettingsReader   — scalar settings, base paths, scope/template defs
  - PathResolver     — entry/note/weekly path & file pattern resolution
  - TemplateProvider — header/inline/time template resolution

Pattern types + defaults moved to config/patterns.ts. TemplateService
(the prior resolver) is split into PathResolver + TemplateProvider, both
pure (no vscode) and fed settings via the IJournalSettings seam. Behavior
unchanged — method bodies relocated, not rewritten.

Refs #234.
